+++ ssc/cmod_battery.h
@@ -122,8 +122,8 @@ struct batt_variables
 
 //    int batt_computed_series;
 //    int batt_computed_strings;
-    size_t batt_computed_series;
-    size_t batt_computed_strings;
+    double batt_computed_series;
+    double batt_computed_strings;
 
 	double batt_kw;
 	double batt_kwh;
@@ -172,8 +172,9 @@ struct batt_variables
 	size_t inverter_model;
 	double inverter_efficiency;
 	double inverter_paco;
-	size_t inverter_count;
-	double batt_inverter_efficiency_cutoff;
+    //size_t inverter_count;
+    double inverter_count;
+    double batt_inverter_efficiency_cutoff;
 
 	double batt_calendar_q0;
 	double batt_calendar_a;
